blueprint化
